refactor: stackflow config 파일 위치 변경
