+++ src/pages/TimetableSelectionActivity/components/TimetableError.tsx
@@ -1,6 +1,6 @@
 import { TimetableContentLayout } from '@/pages/TimetableSelectionActivity/components/TimetableContentLayout';
 import { TimetableMutationErrorStatus } from '@/pages/TimetableSelectionActivity/type';
-import { useFlow } from '@/stackflow';
+import { useFlow } from '@/stackflow/config';
 
 interface TimetableErrorProps {
   status: TimetableMutationErrorStatus;
